Finished first iteration of PathGenerator and Pathfinding Map elements

+++ include/map/Buffer.h
@@ -4,12 +4,12 @@
 
 class Buffer {
     public:
-        Buffer(std::pair<int, int> upper_left_corner, std::pair<int, int> bottom_right_corner) :
+        Buffer(std::pair<double, double> upper_left_corner, std::pair<double, double> bottom_right_corner) :
         upper_left_corner(upper_left_corner), bottom_right_corner(bottom_right_corner) {
 
         }
 
-        bool in_buffer(std::pair<int, int> position) {
+        bool in_buffer(std::pair<double, double> position) {
             if (position.first >= upper_left_corner.first &&
             position.first <= bottom_right_corner.first &&
             position.second <= upper_left_corner.second &&
@@ -18,5 +18,5 @@ class Buffer {
         }
 
     private:
-        std::pair<int, int> upper_left_corner, bottom_right_corner;
+        std::pair<double, double> upper_left_corner, bottom_right_corner;
 };
